Logout button in navbar - Done
