+++ src/components/navbar/navbar.scss
@@ -1,0 +1,9 @@
+.logout-button-navbar {
+  color: white;
+  float: right;
+}
+
+.logout-button-navbar:hover {
+  color: white;
+  text-decoration: underline;
+}
