+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './index.scss'
-import { Layout } from './components/layout'
+import { Layout } from './components/layout/layout'
 import { BrowserRouter as Router } from 'react-router-dom'
 
 const App = () => (
